// src/scenes/dice/dice-polyhedra.generated.ts
export type Vec3 = readonly [number, number, number];

export type Polyhedron = {
  vertices: Vec3[];
  faces: number[][];
};

export type PolyhedronDie = "d4" | "d8" | "d10" | "d12" | "d20";

export const POLYHEDRA: Record<PolyhedronDie, Polyhedron> = {
  d4: {
    vertices: [
      [23.0940107676, 23.0940107676, 23.0940107676],
      [-23.0940107676, -23.0940107676, 23.0940107676],
      [-23.0940107676, 23.0940107676, -23.0940107676],
      [23.0940107676, -23.0940107676, -23.0940107676],
    ],
    faces: [
      [1, 0, 2],
      [3, 0, 1],
      [2, 0, 3],
      [3, 1, 2],
    ],
  },
  d8: {
    vertices: [
      [40, 0, 0],
      [-40, 0, 0],
      [0, 40, 0],
      [0, -40, 0],
      [0, 0, 40],
      [0, 0, -40],
    ],
    faces: [
      [4, 0, 2],
      [2, 0, 5],
      [3, 0, 4],
      [5, 0, 3],
      [2, 1, 4],
      [5, 1, 2],
      [4, 1, 3],
      [3, 1, 5],
    ],
  },
  d10: {
    vertices: [
      [0, 0, 40],
      [21.466252584, 15.5961454192, 4.22291236],
      [21.466252584, -15.5961454192, 4.22291236],
      [26.533747416, 0, -4.22291236],
      [-8.199378876, 25.2350933817, 4.22291236],
      [8.199378876, 25.2350933817, -4.22291236],
      [-26.533747416, 0, 4.22291236],
      [-21.466252584, 15.5961454192, -4.22291236],
      [-8.199378876, -25.2350933817, 4.22291236],
      [-21.466252584, -15.5961454192, -4.22291236],
      [8.199378876, -25.2350933817, -4.22291236],
      [0, 0, -40],
    ],
    faces: [
      [1, 0, 2, 3],
      [5, 4, 0, 1],
      [2, 0, 8, 10],
      [6, 0, 4, 7],
      [8, 0, 6, 9],
      [11, 5, 1, 3],
      [11, 3, 2, 10],
      [7, 4, 5, 11],
      [9, 6, 7, 11],
      [11, 10, 8, 9],
    ],
  },
  d12: {
    vertices: [
      [14.2728835909, 0, 37.3668943585],
      [-14.2728835909, 0, 37.3668943585],
      [0, 37.3668943585, 14.2728835909],
      [23.0940107676, 23.0940107676, 23.0940107676],
      [-23.0940107676, 23.0940107676, 23.0940107676],
      [0, -37.3668943585, 14.2728835909],
      [23.0940107676, -23.0940107676, 23.0940107676],
      [-23.0940107676, -23.0940107676, 23.0940107676],
      [14.2728835909, 0, -37.3668943585],
      [-14.2728835909, 0, -37.3668943585],
      [0, 37.3668943585, -14.2728835909],
      [23.0940107676, 23.0940107676, -23.0940107676],
      [-23.0940107676, 23.0940107676, -23.0940107676],
      [0, -37.3668943585, -14.2728835909],
      [23.0940107676, -23.0940107676, -23.0940107676],
      [-23.0940107676, -23.0940107676, -23.0940107676],
      [37.3668943585, 14.2728835909, 0],
      [-37.3668943585, 14.2728835909, 0],
      [37.3668943585, -14.2728835909, 0],
      [-37.3668943585, -14.2728835909, 0],
    ],
    faces: [
      [4, 1, 0, 3, 2],
      [5, 6, 0, 1, 7],
      [16, 3, 0, 6, 18],
      [19, 7, 1, 4, 17],
      [11, 10, 2, 3, 16],
      [17, 4, 2, 10, 12],
      [18, 6, 5, 13, 14],
      [15, 13, 5, 7, 19],
      [10, 11, 8, 9, 12],
      [15, 9, 8, 14, 13],
      [18, 14, 8, 11, 16],
      [17, 12, 9, 15, 19],
    ],
  },
  d20: {
    vertices: [
      [0, 21.0292444848, 34.0260323341],
      [0, -21.0292444848, 34.0260323341],
      [0, 21.0292444848, -34.0260323341],
      [0, -21.0292444848, -34.0260323341],
      [21.0292444848, 34.0260323341, 0],
      [-21.0292444848, 34.0260323341, 0],
      [21.0292444848, -34.0260323341, 0],
      [-21.0292444848, -34.0260323341, 0],
      [34.0260323341, 0, 21.0292444848],
      [34.0260323341, 0, -21.0292444848],
      [-34.0260323341, 0, 21.0292444848],
      [-34.0260323341, 0, -21.0292444848],
    ],
    faces: [
      [8, 0, 1],
      [1, 0, 10],
      [5, 0, 4],
      [4, 0, 8],
      [10, 0, 5],
      [6, 1, 7],
      [8, 1, 6],
      [7, 1, 10],
      [3, 2, 9],
      [11, 2, 3],
      [4, 2, 5],
      [9, 2, 4],
      [5, 2, 11],
      [7, 3, 6],
      [6, 3, 9],
      [11, 3, 7],
      [9, 4, 8],
      [10, 5, 11],
      [8, 6, 9],
      [11, 7, 10],
    ],
  },
};